% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele.R
\name{bez_naglowka_1kolumny}
\alias{bez_naglowka_1kolumny}
\title{Funkcje nieeksportowane.}
\usage{
bez_naglowka_1kolumny(x)
}
\arguments{
\item{x}{ramka danych}
}
\value{
ramka danych
}
\description{
Funkcja zamienia pierwszą kolumnę ramki danych na nazwy jej
wierszy, aby przy jej drukowaniu móc uzyskać efekt braku nagłówka pierwszej
(drukowanej) kolumny.
}
