% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etykiety_osi_wykresu.R
\name{dec_przecinek}
\alias{dec_przecinek}
\alias{jako_procent}
\title{Formatowanie etykiet osi wykresu.}
\usage{
dec_przecinek(x)

jako_procent(x)
}
\arguments{
\item{x}{liczba}
}
\description{
Funkcje służa formatowaniu etykiet osi Y na wykresach:
odpowiednio zmieniają znak dziesiętny z '.' na ',' oraz dopisują znak '\%'
za liczbą (również zmieniając znak dziesiętny na ',').
}
