% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_wykresu_procenty.R
\name{do_wykresu_procenty}
\alias{do_wykresu_procenty}
\title{Funkcje nieeksportowane.}
\usage{
do_wykresu_procenty(x)
}
\arguments{
\item{x}{ramka danych w postaci \emph{szerokiej}}
}
\value{
ramka danych w postaci \emph{długiej}
}
\description{
Funkcja przekształca ramkę danych zawierającą zestawienie
wartości wskaźników, które mają być zilustrowane wykresem porównującym
rozkłady wyrażone w procentach, do postaci, w której może być ona użyta jako
argument funkcji \code{\link[ggplot2]{ggplot}}.
}
\details{
Funkcja zakłada, że pierwsza kolumna przekazywanej ramki zawiera
etykiety wartości. Oprócz tej kolumny w specjalny sposób traktuje też kolumny
o nazwach \emph{grupa} i \emph{wskaźniki} (o ile występują w przekazanej
ramce danych). Spośród pozostałych kolumn usuwa te, które nie wyglądają jej
na opisujące rozkłady procentowe. Na podstawie nazw pozostałych kolumn stara
się określić, które z nich opisują daną szkołę, a które grupę porównawczą
(dopuszcza tu nieco oboczności zapisu, niemniej ten dychotomiczny podział
jest jedynym w tej chwili obsługiwanym).
}
