% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele.R
\name{dodaj_wiersz_sumy}
\alias{dodaj_wiersz_sumy}
\title{Funkcje nieeksportowane.}
\usage{
dodaj_wiersz_sumy(x, skoryguj_pr_do_100 = TRUE, tekstSuma = "suma")
}
\arguments{
\item{x}{ramka danych}

\item{skoryguj_pr_do_100}{opcjonalnie wartość logiczna - czy korygować sumy
kolumn, \strong{w nazwach których występuje znak '\%'}, aby wynosiły one 100?
(p. sekcja \emph{Details})}

\item{tekstSuma}{opcjonalnie ciąg znaków, który ma zostać użyty jako
etykieta wiersza sumy}
}
\value{
ramka danych
}
\description{
Funkcja dodaje do ramki danych wiersz zawierający sumy kolumn.
}
\details{
Pierwsza kolumna traktowana jest jako zawierająca etykiety wartości.

Korekta sumowania się kolumn, których nazwy zawierają znak '\%' do
100 - wykonywana domyślnie, ale możliwa do wyłączenia przy pomocy argumentu
\code{skoryguj_pr_do_100} - ma na celu obejście problemu zaokrągleń przy
obliczaniu rozkładów częstości i w związku z tym wykonywana jest tylko
w zakresie niewielkich odstępstw sumy wartości w kolumnie od 100. Jeśli
różnica jest większa od 0,3 (tj. wartość sumy mniejsza niż 97 lub większ niż
103), wartość sumy nie zostanie skorygowana. Nigdy nie są też korygowane
wartości w innych wierszach (niż wiersz sumy).
}
