% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabele.R
\name{drukuj_tabele}
\alias{drukuj_tabele}
\title{Drukowanie tabeli z wartosciami wskaznikow.}
\usage{
drukuj_tabele(
  tabela,
  typ = "html",
  dodajWierszSumy = TRUE,
  zamienZnakNowejLinii = as.character(NA),
  align = "center"
)
}
\arguments{
\item{tabela}{ramka danych zawierająca zestawienie wartości wskaźników}

\item{typ}{"html" (domyślnie) lub "pdf" - format, w jakim ma zostać
zwrócony kod tabeli}

\item{dodajWierszSumy}{opcjonalnie wartość logiczna - czy dodawać do tabeli
wiersz sumy? (jeśli tak, wszystkie kolumny poza pierwszą powinny zawierać
liczby)}

\item{zamienZnakNowejLinii}{opcjonalnie ciąg znaków, który ma zostać
podstawiony w treści tabeli, ale \strong{nie w wierszu nagłówka}, zamiast
oznaczenia nowej linii (tj. zamiast ciągu '\\n')}

\item{align}{"center" (domyślnie) "left" lub "right" - wyrównanie tabeli}
}
\value{
ciąg znaków zawierający kod w wybranym formacie opisujący tabelę
}
\description{
Funkcja przygotowuję tabelę z wartościami wskaźników do
wydrukowania i zwraca opisujący ją kod w formacie HTML lub LaTeX (PDF).
}
