% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj_raporty.R
\name{generuj_raport}
\alias{generuj_raport}
\title{Funkcje nieeksportowane.}
\usage{
generuj_raport(szablonZeSciezka, nazwaPliku, parametry, wskaznikiGrupa,
  wskaznikiGrPor)
}
\arguments{
\item{szablonZeSciezka}{ciąg znaków - ścieżka do pliku z szablonem raportu}

\item{nazwaPliku}{ciąg znaków - nazwa pliku raportu, któy ma zostać utworzony}

\item{parametry}{lista parametrów przekazywanych do szablonu}

\item{wskaznikiGrupa}{ramka danych zawierająca tylko jeden wiersz, ze
wskaznikami grupy, dla której ma zostać utworzony raport}

\item{wskaznikiGrPor}{opcjonalnie ramka danych ze wskaźnikami grup
porównawczych, do wykorzystania przy tworzeniu raportu}
}
\value{
obiekt zwracany przez funkcję \code{\link[knitr]{knit_meta}}
zawierający metdane dotyczące utworzonego raportu
}
\description{
Funkcja pozwala oddzielić środowisko, w którym odbywa się
generowanie pojedynczego raportu od środowiska funkcji
\code{\link{generuj_raporty}}.
}
