% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komunikat_o_liczebnosciach.R
\name{komunikat_o_liczebnosci}
\alias{komunikat_o_liczebnosci}
\title{Drukowanie informacji o liczbie absolwentow, na podsatwie ktorych
obliczono wartosci wskaznika.}
\usage{
komunikat_o_liczebnosci(wskaznik, progLiczebnosci, opis = "")
}
\arguments{
\item{wskaznik}{obiekt (lista) z wartościami wskaźnika w danej grupie}

\item{progLiczebnosci}{liczba - próg liczebności, poniżej którego powinien
zostać zwrócony komunikat o niemożliwości zaprezentowania wskaźnika}

\item{opis}{opcjonalnie ciąg znaków opisujący grupę absolwentów, dla której
obliczony został wskaźnik (np. ", którzy mieli pracę w rok po ukończeniu
szkoły")}
}
\value{
ciąg znaków
}
\description{
Funkcja przygotowuje komunikat o tym, jaka liczba absolwentów
została uwzględniona przy obliczaniu wartości wskaźnika lub komunikat o tym,
że wskaźnik nie może zostać zaprezentowany ze względu na zbyt małą liczbę
badanych.
}
