% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liczba_abs_wojewodztwa}
\alias{liczba_abs_wojewodztwa}
\title{Liczba absolwentów w danym typie szkoły i województwie}
\format{
Ramka danych
}
\usage{
liczba_abs_wojewodztwa
}
\description{
Obiekt zawierający rozkład liczby absolwentów w danym typie
szkoły i województwie
}
\keyword{datasets}
