% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/szablony.R
\name{okresl_lokalizacje_szablonow}
\alias{okresl_lokalizacje_szablonow}
\title{Funkcje nieeksportowane.}
\usage{
okresl_lokalizacje_szablonow()
}
\value{
ścieżka do folderu
}
\description{
Funkcja określa, gdzie znajdują się szablony raportów
(w zależności od tego, czy pakiet jest testowany, czy już zainstalowany
znajdują się one albo w folderze 'inst/szablony_raportow' albo w folderze
'szablony_raportow/' głównego folderu pakietu) i zwraca ścieżkę do
odpowiedniego folderu.
}
