% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polska}
\alias{polska}
\title{Przykladowe wskazniki obliczone na poziomie typów szkół}
\format{
Ramka danych z 3 wierszami - typy szkół
}
\usage{
polska
}
\description{
Obiekt ze wskaźnikami dla typów szkół w całej Polsce.
}
\keyword{datasets}
