% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskaznikiSzkGrPor}
\alias{wskaznikiSzkGrPor}
\title{Przykladowe wskazniki obliczone na poziomie typow szkol.}
\format{
Ramka danych z 20 wierszami (szkołami) i 75 kolumnami (wskaznikami).
}
\usage{
wskaznikiSzkGrPor
}
\description{
Wskaźniki opisujące grupy porównawcze (dla wartości wskaźników
poszczególnych szkół) odpowiadające grupom z ramki danych
\code{wskaznikiSZK}, obliczone na poziomie typów szkół z wykorzystaniem
danych z I rundy monitoringu losów absolwentów szkół zawodowych zrealizowanej
w ramach projektu MLEZAiMD. Do wykorzystania przy generowaniu przykładowych
raportów.
}
\keyword{datasets}
