% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskaznikiTypSzk}
\alias{wskaznikiTypSzk}
\title{Przykladowe wskazniki obliczone na poziomie typow szkol.}
\format{Ramka danych z 3 wierszami (szkołami) i 68 kolumnami (wskaznikami).}
\usage{
wskaznikiTypSzk
}
\description{
Wskaźniki opisujące grupy odniesienia (dla wartości wskaźników
poszczególnych szkół), obliczone na poziomie typów szkół z wykorzystaniem
danych z I rundy monitoringu losów absolwentów szkół zawodowych zrealizowanej
w ramach projektu MLEZAiMD. Do wykorzystania przy generowaniu przykładowych
raportów.
}
\keyword{datasets}
