% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskaznikiZawPow}
\alias{wskaznikiZawPow}
\title{Przykladowe wskazniki obliczone na poziomie powiato-zawodow.}
\format{
Ramka danych z 5 wierszami (powiato-zawodami) i 72 kolumnami (wskaznikami).
}
\usage{
wskaznikiZawPow
}
\description{
Zanonimizowane wskaźniki dla kilku powiato-zawodów z I rundy monitoringu
losów absolwentów szkół zawodowych zrealizowanej w ramach projektu MLEZAiMD,
na podstawie których można wygenerować przykładowe raporty.
}
\keyword{datasets}
