% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wskaznikiZawWoj}
\alias{wskaznikiZawWoj}
\title{Przykladowe wskazniki obliczone na poziomie wojewodztw.}
\format{Ramka danych z 5 wierszami (województwo-zawodami) i 68 kolumnami (wskaznikami).}
\usage{
wskaznikiZawWoj
}
\description{
Wskaźniki opisujące grupy odniesienia (dla wartości wskaźników
poszczególnych powiato-zawodów), obliczone na poziomie województwo-zawodów z wykorzystaniem
danych z I rundy monitoringu losów absolwentów szkół zawodowych zrealizowanej
w ramach projektu MLEZAiMD. Do wykorzystania przy generowaniu przykładowych
raportów.
}
\keyword{datasets}
