% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_pudelkowy.R
\name{wykres_pudelkowy}
\alias{wykres_pudelkowy}
\title{Wykres pudelkowy wskaznika.}
\usage{
wykres_pudelkowy(tabela, zmienna, etykietaY,
  funkcjaFormatujacaY = MLASZraporty::dec_przecinek)
}
\arguments{
\item{tabela}{ramka danych zawierająca zestawienie wartości wskaźników}

\item{zmienna}{wyrażenie lub ciąg znaków - nazwa kolumny w przekazywanej
ramce danych, w której podane są opisy parametrów, których wartości zawiera
zestawienie (tj. tej kolumny, na podstawie której można się zorientować,
w którym wierszu opisane są wartości 1. kwartyle, w którym mediany itd.)}

\item{etykietaY}{ciąg znaków - etykieta osi Y wykresu}

\item{funkcjaFormatujacaY}{opcjonalnie funkcja, która ma zostać użyta do
formatowania etykiet wartości osi Y wykresu}
}
\value{
wykres ggplot2
}
\description{
Funkcja zwraca wykres pudełkowy ilustrujący wartości wskaźników,
opisywanych przez 1. i 3. kwartyl, medianę oraz średnią.
}
