% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_rozklad_liniowy.R
\name{wykres_rozklad_liniowy}
\alias{wykres_rozklad_liniowy}
\title{Wykres liniowy wskaznika.}
\usage{
wykres_rozklad_liniowy(
  tabela,
  facet = NULL,
  etykietaX = NULL,
  etykietaY = "częstość"
)
}
\arguments{
\item{tabela}{ramka danych zawierająca zestawienie wartości wskaźników}

\item{facet}{opcjonalnie wyrażenie lub ciąg znaków - nazwa kolumny
w przekazywanej ramce danych, na podstawie wartości której dane zostaną
podzielone pomiędzy \emph{segmenty} wykresu}

\item{etykietaX}{opcjonalnie ciąg znaków - etykieta osi X wykresu}

\item{etykietaY}{opcjonalnie ciąg znaków - etykieta osi Y wykresu}
}
\value{
wykres ggplot2
}
\description{
Funkcja zwraca wykres liniowy ilustrujący rozkład wartości
wskaźników.
}
\details{
Funkcja zakłada, że przekazywana tabela zawiera kolumnę o nazwie
\code{etykieta} lub \code{wskaznik}, która opisuje etykiety wartosci
ilustrowanych rozkładów (tj. wartości, które wystąpią na osi X wykresu).
}
