% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_slupkowe.R
\name{wykres_slupkowy_procenty}
\alias{wykres_slupkowy_procenty}
\title{Wykres slupkowy wskaznika mogacego sumowac sie do wiecej niz 100\%.}
\usage{
wykres_slupkowy_procenty(tabela, zmienna)
}
\arguments{
\item{tabela}{ramka danych zawierająca zestawienie wartości wskaźników}

\item{zmienna}{wyrażenie lub ciąg znaków - nazwa kolumny w przekazywanej
ramce danych, w której podane są etykiety wartości wskaźników, których
rozkłady mają zostać ze sobą porównane}
}
\value{
wykres ggplot2
}
\description{
Funkcja zwraca wykres słupkowy ilustrujący wartości wskaźników,
w którym słupki nie są \emph{układane jeden na drugim}, tylko obok siebie.
Jest ona odpowiednia do ilustrowania wartości wskaźników procentowych,
których suma wartości może przekraczać 100\%.
}
