% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy_slupkowe.R
\name{wykres_slupkowy_skumulowany_procenty}
\alias{wykres_slupkowy_skumulowany_procenty}
\title{Wykres slupkowy skumulowany wskaznika.}
\usage{
wykres_slupkowy_skumulowany_procenty(tabela, zmienna, x = "klucz",
  facet = NULL)
}
\arguments{
\item{tabela}{ramka danych zawierająca zestawienie wartości wskaźników}

\item{zmienna}{wyrażenie lub ciąg znaków - nazwa kolumny w przekazywanej
ramce danych, w której podane są etykiety wartości wskaźników (tj. kolumna
wyznaczająca podział w ramach słupków)}

\item{x}{wyrażenie lub ciąg znaków - nazwa kolumny w przekazywanej ramce
danych, której wartości będą ilustrowane na osi X wykresu (tj. kolumna
wyznaczająca podział na słupki)}

\item{facet}{opcjonalnie wyrażenie lub ciąg znaków - nazwa kolumny
w przekazywanej ramce danych, na podstawie wartości której dane zostaną
podzielone pomiędzy \emph{segmenty} wykresu}
}
\value{
wykres ggplot2
}
\description{
Funkcja zwraca wykres słupkowy skumulowany ilustrujący
wartości wskaźników.
}
