% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/szablony.R
\name{wypisz_dostepne_szablony}
\alias{wypisz_dostepne_szablony}
\title{Lista szablonow raportow dostepnych w ramach pakietu.}
\usage{
wypisz_dostepne_szablony()
}
\value{
wektor tekstowy (niewidocznie)
}
\description{
Funkcja pokazuje nazwy szablonów raportów, które są dostępne
w ramach pakietu.
}
