% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_wykresu_pudelkowego.R
\name{do_wykresu_pudelkowego}
\alias{do_wykresu_pudelkowego}
\title{Funkcje nieeksportowane.}
\usage{
do_wykresu_pudelkowego(x)
}
\arguments{
\item{x}{ramka danych w postaci \emph{szerokiej}}
}
\value{
ramka danych w postaci \emph{długiej}
}
\description{
Funkcja przekształca ramkę danych zawierającą zestawienie
wartości wskaźników, które mają być zilustrowane wykresem pudełkowym do
postaci, w której może być ona użyta jako argument funkcji
\code{\link[ggplot2]{ggplot}}.
}
